chore(cli): Remove legacy references
- Removes V1 migration steps
- Removes references to "legacy" types and variables
